Pause Menu (opens on esc.)
Added a Pause Menu. The menu opens and closes when you press escape.

The Main button opens up the main menu, the Resume button continues the game, and the Quit button doesn't really do anything right now. There's a quit function that hopefully closes the game when in the launcher.